src/backend/requirements.txt
mlflow==3.3.1
cloudpickle==3.1.1
numpy==1.26.4
pandas==2.3.2
psutil==7.0.0
pyarrow==21.0.0
scikit-learn==1.7.1
scikit-optimize==0.10.2
scipy==1.16.1
xgboost==3.0.4
python-dotenv==1.1.1
fastapi==0.116.1
uvicorn==0.35.0
nest-asyncio==1.6.0
mlflow[databricks]
shap==0.48.0